+++ R/pp_check.R
@@ -149,7 +149,7 @@ pp_check = function(
 #' @param yrep S X N matrix of predicted responses
 #' @param observations (required for loo-type plots) Original data-row indices.
 #' @param ... Arguments passed to `bayesplot::ppc_type(y, yrep, ...)`
-#' @return A `ggplot2` object returned by `tidybayes::ppc_*(y, yrep, ...)`.
+#' @return A `ggplot2` object returned by `bayesplot::ppc_*(y, yrep, ...)`.
 #' @return A string
 #' @encoding UTF-8
 #' @author Jonas Kristoffer Lindeløv \email{jonas@@lindeloev.dk}